Small corrections on classroom and lecture repositories.

+++ database/src/main/java/nl/tudelft/unischeduler/database/repositories/ClassroomRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ClassroomRepository extends JpaRepository<Classroom, Long> {
+public interface ClassroomRepository extends JpaRepository<Classroom, Integer> {
     List<Classroom> findAll();
 }
